backend/requirements.txt
# FastAPI and web framework
fastapi[standard]>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0

# Database
sqlalchemy>=2.0.30,<2.1.0
aiosqlite>=0.20.0,<0.21.0
alembic>=1.14.0,<1.15.0

# Redis caching
redis>=5.2.0,<6.0.0
hiredis>=3.0.0,<4.0.0

# Vector database
pymilvus>=2.4.0,<2.5.0

# Data validation and serialization
pydantic>=2.10.0,<2.11.0
pydantic-settings>=2.6.0,<2.7.0

# Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Async HTTP client
httpx>=0.28.0,<0.29.0

# Background tasks
celery>=5.4.0,<6.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Logging and monitoring
structlog>=24.4.0,<25.0.0

# Testing (development)
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<0.25.0
pytest-cov>=6.0.0,<7.0.0

# Type checking (development)
mypy>=1.13.0,<2.0.0
types-redis>=4.6.0